requirements.txt
# Version 
pytesseract==0.3.10
Pillow==9.5.0 
pyodbc==5.0.1
pandas==2.0.0

